Refactor dialogue methods into Dialogue class
Moved all dialogue-related methods from RedGirls.java to a new charisma.Dialogue class for better separation of concerns and code organization.
Updated RedGirls to use the new Dialogue class.
Also removed unnecessary exception from Parser.parseInput signature.
