+++ src/main/java/charisma/Dialogue.java
@@ -1,0 +1,55 @@
+package charisma;
+
+public class Dialogue {
+    public static String toRedGirlsString(String dialogue) {
+        String darkRed = "\033[38;5;88m";
+        String reset = "\033[0m";
+        return darkRed + dialogue + reset;
+    }
+
+    public static void redGirlsPrint(String dialogue) {
+        System.out.println("\n" + toRedGirlsString(dialogue));
+    }
+
+    public static void printBootSequence() {
+        String[] bootLines = {
+                "\033[1;36mBOOT LOADING - BOOTING SYSTEM\033[0m",
+                "Commencing System Check",
+                "Memory Unit: \033[32mGreen\033[0m",
+                "Initializing Tactics Log",
+                "Loading Geographic Data",
+                "Vitals: \033[32mGreen\033[0m",
+                "Remaining MP: \033[32m100%\033[0m",
+                "Black Box Temperature: \033[32mNormal\033[0m",
+                "Black Box Internal Pressure: \033[32mNormal\033[0m",
+                "Activating IFF",
+                "Activating FCS",
+                "Initializing Pod Connection",
+                "Launching DBU Setup",
+                "Activating Inertia Control System",
+                "Activating Environmental Sensors",
+                "Equipment Authentication: \033[32mComplete\033[0m",
+                "Equipment Status: \033[32mGreen\033[0m",
+                "\033[1;32mAll Systems Green\033[0m",
+                "\033[1;32mCombat Preparations Complete\033[0m"
+        };
+
+        for (String line : bootLines) {
+            System.out.println(line);
+            try {
+                Thread.sleep(200);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+    }
+
+    public static void printGreeting() {
+        redGirlsPrint("I perceive the fragments of your thoughts…\n" +
+                "How intriguing. Shall we converse?");
+    }
+
+    public static void printFarewell() {
+        redGirlsPrint("Our exchange concludes. Your thoughts linger...as do mine.");
+    }
+}
